requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
google-cloud-logging
google-cloud-storage
pandas
plotly
protobuf
streamlit>=1.46
altair>=5.0.0
PyYAML>=5.4
python-dotenv
pysftp
paramiko<3.0